backend/requirements.txt
Django>=5.0
djangorestframework
psycopg2-binary
gunicorn
django-environ
django-cors-headers
djangorestframework-simplejwt
requests
beautifulsoup4
plaid-python
yfinance==0.2.40
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
scipy==1.14.1
python-dateutil==2.8.2